refactor(internal): move HTTP and WebSocket implementation headers to internal
Move HTTP client/server and WebSocket client/server implementation headers
from include/kcenon/network/http/ to src/internal/http/ to reduce public
API surface.

Changes:
- Create internal headers in src/internal/http/
- Convert public headers to redirect stubs with deprecation warnings
- Update facade implementations to use internal headers
- Update .cpp implementations to use internal headers directly

Part of Epic #577 - Apply Facade pattern to reduce protocol header complexity.

Closes #636
