+++ src/http/http_client.cpp
@@ -30,7 +30,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
-#include "kcenon/network/http/http_client.h"
+#include "internal/http/http_client.h"
 #include <condition_variable>
 #include <mutex>
 #include <regex>
