+++ src/http/http_server.cpp
@@ -30,7 +30,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
-#include "kcenon/network/http/http_server.h"
+#include "internal/http/http_server.h"
 #include "kcenon/network/session/messaging_session.h"
 #include "kcenon/network/utils/compression_pipeline.h"
 #include <algorithm>
