+++ src/http/websocket_client.cpp
@@ -30,7 +30,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
-#include "kcenon/network/http/websocket_client.h"
+#include "internal/http/websocket_client.h"
 
 #include "internal/core/network_context.h"
 #include "internal/tcp/tcp_socket.h"
